Minimal changes needed to install Roon 2.0 with Wine 7
Given Wine 7 is installed properly, i.e. following standard instructions
such as: https://wine.htmlvalidator.com/install-wine-on-pop_os-22.04.html then
this is the fewest changes necessary to get Roon 2.0 installing properly.

Sources for these changes:
* https://community.roonlabs.com/t/roon-2-0-and-linux-wine/213628/11

RoPieee/roon-on-wine#27 may also be used but this is a more minimal
set of changes so may be preferable.