Reduce CPU usage by suppressing Wine warnings
This commit addresses an issue observed on Linux Mint 21.1 when running Roon with Wine 8.0. The terminal window gets flooded with warnings like "fixme:manipulation:update_manager_Update", causing CPU utilization to spike to 100%. It is currently unclear if this issue affects other Linux distributions as well.

To improve performance and reduce CPU usage, this fix suppresses the aforementioned warnings. As a result, CPU utilization should return to normal levels, providing a better experience.

Note: This fix is focused on suppressing verbose warning messages that were causing performance issues and does not impact the core functionality of the application or Wine itself.